tests/baseline/protocol-1.2/inspector/console/command-line-api-getEventListeners-expected.txt

CONSOLE ERROR: line 26: Uncaught SyntaxError: Unexpected identifier
Tests getEventListeners() method of console command line API.

 
[page] - inner -
[page] keydown: {
[page]     0: {
[page]         listener: function listener1() { }
[page]         passive: false
[page]         remove: function remove() { [Command Line API] }
[page]         type: "keydown"
[page]         useCapture: false
[page]     }
[page]     1: {
[page]         listener: function listener2() { }
[page]         passive: false
[page]         remove: function remove() { [Command Line API] }
[page]         type: "keydown"
[page]         useCapture: true
[page]     }
[page] }
[page] wheel: {
[page]     0: {
[page]         listener: function listener2() { }
[page]         passive: true
[page]         remove: function remove() { [Command Line API] }
[page]         type: "wheel"
[page]         useCapture: false
[page]     }
[page] }
[page] - inner after a removal -
[page] keydown: {
[page]     0: {
[page]         listener: function listener2() { }
[page]         passive: false
[page]         remove: function remove() { [Command Line API] }
[page]         type: "keydown"
[page]         useCapture: true
[page]     }
[page] }
[page] - outer -
[page] keydown: {
[page]     0: {
[page]         listener: function listener2() { }
[page]         passive: false
[page]         remove: function remove() { [Command Line API] }
[page]         type: "keydown"
[page]         useCapture: true
[page]     }
[page] }
[page] keyup: {
[page]     0: {
[page]         listener: function listener2() { }
[page]         passive: false
[page]         remove: function remove() { [Command Line API] }
[page]         type: "keyup"
[page]         useCapture: false
[page]     }
[page] }
[page] mousedown: {
[page]     0: {
[page]         listener: function listener2() { }
[page]         passive: false
[page]         remove: function remove() { [Command Line API] }
[page]         type: "mousedown"
[page]         useCapture: true
[page]     }
[page] }
[page] mousemove: {
[page]     0: {
[page]         listener: function listener1() { }
[page]         passive: false
[page]         remove: function remove() { [Command Line API] }
[page]         type: "mousemove"
[page]         useCapture: false
[page]     }
[page] }
[page] - attribute event listeners -
[page] click: {
[page]     0: {
[page]         listener: function onclick(event) { alert(1) }
[page]         passive: false
[page]         remove: function remove() { [Command Line API] }
[page]         type: "click"
[page]         useCapture: false
[page]     }
[page] }
[page] mouseover: {
[page]     0: {
[page]         listener: function onmouseover(event) { listener2() }
[page]         passive: false
[page]         remove: function remove() { [Command Line API] }
[page]         type: "mouseover"
[page]         useCapture: false
[page]     }
[page] }
[page] - window -
[page] load: {
[page]     0: {
[page]         listener: function onload(event) { runTest() }
[page]         passive: false
[page]         remove: function remove() { [Command Line API] }
[page]         type: "load"
[page]         useCapture: false
[page]     }
[page] }
[page] popstate: {
[page]     0: {
[page]         listener: function listener1() { }
[page]         passive: false
[page]         remove: function remove() { [Command Line API] }
[page]         type: "popstate"
[page]         useCapture: false
[page]     }
[page] }
[page] - empty -
[page] - invalid -
[page] - object -
[page] object
[page] - null -
[page] object
[page] - undefined -
[page] object
